# Faster findall for bitarrays