@chethega 2018-11-06
Sorry for taking so long to properly respond to this.

Very cool. Your solution widely surpasses my expectation when I issued the challenge on discourse, I did not expect that this can get so fast!