@maxbennedich 2018-11-07
Thanks for that. I added a few tests now. What's the next step here? Would anyone like to take a stab at reviewing the code?