@mbauman 2018-11-07
I tried a few other sizes with your benchmarking script trying to assess the worst case scenarios… and even then this is performing spectacularly. I had to bend over backwards to find _anything_ that's remotely a regression, and even then only in a few circumstances!

```
       size      | selected |  old time  |   per idx  |  cycles |  new time  |   per idx  |  cycles | speedup
    1 x 201 x 10 |    0.1 % |    2.91 μs | 1453.33 ns | 3197.33 |    0.71 μs |  352.79 ns |  776.13 |  4.12 x
    1 x 201 x 10 |    0.7 % |    2.98 μs |  212.87 ns |  468.32 |    0.89 μs |   63.61 ns |  139.94 |  3.35 x
    1 x 201 x 10 |    5.2 % |    3.51 μs |   33.77 ns |   74.30 |    1.29 μs |   12.43 ns |   27.34 |  2.72 x
    1 x 201 x 10 |   20.0 % |    4.26 μs |   10.57 ns |   23.24 |    1.53 μs |    3.81 ns |    8.37 |  2.78 x
    1 x 201 x 10 |   50.8 % |    4.38 μs |    4.29 ns |    9.44 |    2.58 μs |    2.52 ns |    5.55 |  1.70 x
    1 x 201 x 10 |   79.9 % |    4.54 μs |    2.83 ns |    6.22 |    3.62 μs |    2.25 ns |    4.96 |  1.25 x
    1 x 201 x 10 |   98.9 % |    4.98 μs |    2.51 ns |    5.52 |    4.45 μs |    2.24 ns |    4.93 |  1.12 x
    1 x 201 x 10 |  100.0 % |    4.93 μs |    2.45 ns |    5.40 |    3.89 μs |    1.93 ns |    4.25 |  1.27 x
    2 x 3 x 1000 |    0.2 % |   11.98 μs | 1089.09 ns | 2396.00 |    1.64 μs |  148.96 ns |  327.72 |  7.31 x
    2 x 3 x 1000 |    1.0 % |   12.14 μs |  209.26 ns |  460.37 |    1.94 μs |   33.39 ns |   73.45 |  6.27 x
    2 x 3 x 1000 |    4.8 % |   12.92 μs |   45.33 ns |   99.73 |    2.30 μs |    8.07 ns |   17.75 |  5.62 x
    2 x 3 x 1000 |   19.7 % |   16.15 μs |   13.64 ns |   30.01 |    4.45 μs |    3.75 ns |    8.26 |  3.63 x
    2 x 3 x 1000 |   50.9 % |   24.21 μs |    7.93 ns |   17.45 |   11.66 μs |    3.82 ns |    8.40 |  2.08 x
    2 x 3 x 1000 |   80.6 % |   13.94 μs |    2.88 ns |    6.34 |   13.85 μs |    2.86 ns |    6.30 |  1.01 x
    2 x 3 x 1000 |   99.1 % |   12.67 μs |    2.13 ns |    4.69 |   12.23 μs |    2.06 ns |    4.53 |  1.04 x
    2 x 3 x 1000 |  100.0 % |   12.20 μs |    2.03 ns |    4.47 |   10.16 μs |    1.69 ns |    3.73 |  1.20 x
  1 x 1 x 100000 |    0.1 % |  140.78 μs | 1366.84 ns | 3007.06 |   72.82 μs |  707.01 ns | 1555.42 |  1.93 x
  1 x 1 x 100000 |    1.0 % |  146.35 μs |  145.33 ns |  319.73 |   93.78 μs |   93.12 ns |  204.87 |  1.56 x
  1 x 1 x 100000 |    5.0 % |  190.70 μs |   38.39 ns |   84.47 |  156.51 μs |   31.51 ns |   69.32 |  1.22 x
  1 x 1 x 100000 |   20.1 % |  352.04 μs |   17.51 ns |   38.51 |  305.54 μs |   15.19 ns |   33.43 |  1.15 x
  1 x 1 x 100000 |   50.0 % |  503.80 μs |   10.07 ns |   22.15 |  503.27 μs |   10.06 ns |   22.13 |  1.00 x
  1 x 1 x 100000 |   80.0 % |  338.91 μs |    4.24 ns |    9.32 |  355.38 μs |    4.44 ns |    9.77 |  0.95 x
  1 x 1 x 100000 |   99.0 % |  249.24 μs |    2.52 ns |    5.54 |  245.93 μs |    2.49 ns |    5.47 |  1.01 x
  1 x 1 x 100000 |  100.0 % |  252.11 μs |    2.52 ns |    5.55 |  218.66 μs |    2.19 ns |    4.81 |  1.15 x
  2 x 1 x 100000 |    0.1 % |  394.49 μs | 1915.01 ns | 4213.03 |   76.54 μs |  371.56 ns |  817.44 |  5.15 x
  2 x 1 x 100000 |    1.0 % |  406.35 μs |  204.61 ns |  450.13 |  117.61 μs |   59.22 ns |  130.28 |  3.46 x
  2 x 1 x 100000 |    4.9 % |  451.07 μs |   45.79 ns |  100.73 |  224.80 μs |   22.82 ns |   50.20 |  2.01 x
  2 x 1 x 100000 |   20.1 % |  667.98 μs |   16.65 ns |   36.63 |  494.97 μs |   12.34 ns |   27.14 |  1.35 x
  2 x 1 x 100000 |   50.1 % |  945.96 μs |    9.44 ns |   20.77 |  643.56 μs |    6.42 ns |   14.13 |  1.47 x
  2 x 1 x 100000 |   80.1 % |  647.65 μs |    4.04 ns |    8.90 |  625.24 μs |    3.90 ns |    8.59 |  1.04 x
  2 x 1 x 100000 |   99.0 % |  429.32 μs |    2.17 ns |    4.77 |  481.60 μs |    2.43 ns |    5.35 |  0.89 x
  2 x 1 x 100000 |  100.0 % |  471.20 μs |    2.36 ns |    5.18 |  419.94 μs |    2.10 ns |    4.62 |  1.12 x
1 x 1 x 2 x 1 x 10000 |    0.1 % |   43.43 μs | 2068.14 ns | 4549.91 |   21.44 μs | 1021.10 ns | 2246.41 |  2.03 x
1 x 1 x 2 x 1 x 10000 |    1.1 % |   45.44 μs |  212.32 ns |  467.10 |   28.08 μs |  131.22 ns |  288.68 |  1.62 x
1 x 1 x 2 x 1 x 10000 |    5.1 % |   52.11 μs |   51.59 ns |  113.50 |   44.14 μs |   43.70 ns |   96.14 |  1.18 x
1 x 1 x 2 x 1 x 10000 |   19.9 % |   75.65 μs |   19.05 ns |   41.92 |   79.29 μs |   19.97 ns |   43.94 |  0.95 x
1 x 1 x 2 x 1 x 10000 |   49.7 % |  113.96 μs |   11.46 ns |   25.20 |  126.28 μs |   12.69 ns |   27.93 |  0.90 x
1 x 1 x 2 x 1 x 10000 |   80.1 % |   91.50 μs |    5.71 ns |   12.56 |  110.59 μs |    6.90 ns |   15.18 |  0.83 x
1 x 1 x 2 x 1 x 10000 |   99.0 % |   84.66 μs |    4.28 ns |    9.41 |  105.26 μs |    5.32 ns |   11.70 |  0.80 x
1 x 1 x 2 x 1 x 10000 |  100.0 % |   85.04 μs |    4.25 ns |    9.35 |   53.54 μs |    2.68 ns |    5.89 |  1.59 x
```