@StefanKarpinski 2018-11-08
What remains to be done here? This PR seems to be in good shape. Is it time to merge?