@chethega 2018-11-08
>Is it time to merge?

Nanosoldier to catch weird surprises, and either delight in the nice new number or add "benchmarks beneficial" tag before merging?