@KristofferC 2018-11-08
@nanosoldier `runbenchmarks(ALL, vs = ":master")`