@maxbennedich 2018-11-10
What happened to the nanosoldier run? Looking at BaseBenchmarkReports, the "daily" report has not been produced since Nov 3, so perhaps the service is having trouble? (Although there was a run completing 3 hours ago.)