@nanosoldier 2018-11-10
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29888#issuecomment-437587647) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/dfc4ad82050db3b44d4e8914b985f812f16c39ec/7f94a8c_vs_a3599e9/report.md). cc @ararslan